# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateProjectRequest:
    properties:
      description:
        example: описание проекта
        type: string
      name:
        example: Написать API
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Произошла ошибка
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      message:
        example: Операция выполнена успешно
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Support Team
  description: API для управления задачами, проектами и комментариями
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SaaS
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Возвращает список проектов авторизованного пользователя и их описание.
        Требуются права просмотра проекта.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Получить список проектов
      tags:
      - project
    post:
      consumes:
      - application/json
      description: Создаёт новую запись в таблице проектов.
      parameters:
      - description: название и описание проекта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать проект
      tags:
      - project
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись из таблицы проектов. Требуются права редактора проекта.
      parameters:
      - description: id проекта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить проект
      tags:
      - project
    get:
      consumes:
      - application/json
      description: Возвращает проект и его описание. Требуются права просмотра проекта.
      parameters:
      - description: id проекта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить конкретный проект
      tags:
      - project
    patch:
      consumes:
      - application/json
      description: Изменяет данные проекта на данные из тела запроса. Требуются права
        редактора проекта.
      parameters:
      - description: id проекта
        in: path
        name: id
        required: true
        type: integer
      - description: название и описание проекта для редактирования
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Изменить проект
      tags:
      - project
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer {token}" (без кавычек)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"